refactor(ast_visit): move custom visitors for UTF8 to UTF16 offset conversion (#13237)
Pure refactor.

Previously there was a big block of static code for custom visitors included in the codegen for `impl VisitMut for Utf8ToUtf16Converter`.

Move this code out of the codegen and into its own `.rs` file. The advantage is that you now get type hints from Rust Analyser when exploring/modifying this code. This change also simplifies the codegen a lot.

Similar to #13236, the intent is to make the converter code a bit easier to follow, to hopefully enable external contributors to make progress on adding support for `loc` in ESTree AST (#10307).