feat(linter): add jest/padding-around-test-blocks rule (#12985)
Adds the
[`jest/padding-around-test-blocks`](https://github.com/jest-community/eslint-plugin-jest/blob/v29.0.1/docs/rules/padding-around-test-blocks.md)
rule. The jest plugin opts for abstracting away all the
`padding-around-x` rules with a helper, but for my purposes, I only
needed this specific rule. My opinion is that we can investigate
generalizing this rule to the other padding plugin as a followup, if
there is pull for it.

Works towards https://github.com/oxc-project/oxc/issues/492