feat(minifier): remove unused assignments for vars (#13231)
I found that the temporary variable are kept. This change would remove them.
Variables that were declared with `var` did not have `SymbolValue` populated and that caused those variables to bail out at this line.
https://github.com/oxc-project/oxc/blob/2141c18540803218889aa3c3f1bca5f51cdb5daa/crates/oxc_minifier/src/peephole/remove_unused_expression.rs#L624-L626

I think TDZ needs to be considered when inlining the values, but does not affect removing the unused assignments.