refactor(ast_visit): `Utf8ToUtf16` record end offset of multi-byte chars (#13339)
`Utf8ToUtf16` converter creates a table of translations from UTF-8 offset to UTF-16 offset.

Previously it recorded UTF-8 offsets as the start of the multi-byte character + 1. This PR changes that to record the offset of the *end* of character.

This costs a couple more instructions when adding records to the table, but makes no difference to the amount of work involved in converting offsets.

The advantages are:

1. Reduces complexity due to not having to handle edge cases where arithmetic overflow was possible previously.
2. Makes it possible to use the same table to convert back from UTF-16 to UTF-8 (next PR in this stack).

As Unicode chars are rare, and the extra cost is small anyway, I think this trade-off is worth it.