feat(linter): convert spans to UTF-16 in JS plugins (#13344)
The AST on JS side in linter JS plugins previously had `start` and `end` as UTF-8 offsets.

Convert the AST to UTF-16 offsets prior to transferring it to JS, and convert offsets in any errors/warnings returned from JS side back to UTF-8 before creating `OxcDiagnostic`s from them.

The tricky part is that it's not possible to mutate the AST (convert offsets to UTF-16) while `Semantic` is alive, because you can't obtain a `&mut Program` due to aliasing rules. Work around this with some pointer juggling and by dropping `Semantic` before creating a `&mut Program`.

This is a bit sketchy, but is not (I believe) UB. Hopefully we can remove the need for this workaround later on by performing offset conversion during serialization / raw transfer deserialization, and so avoid the need to mutate the AST.