chore(codegen): skip stacktrace tests on big endian test workflow (#13382)
This PR tries to fix the following test failure.
https://github.com/oxc-project/oxc/actions/runs/17313194902/job/49151067079#step:5:325

This failure is probably happening because node binary is missing in the docker image that the test is running. While we can add the node binary in the docker container, I think we can skip the test as I guess there aren't any endian-specific code.