perf(allocator): remove bounds checks in `CloneIn` for `Vec` (#13506)
Improve the performance of `CloneIn` for `Vec`.

Previously we used `Vec::from_iter_in`, which is pretty inefficient. It performs a bounds check on every turn of the loop, which is unnecessary because we know the length of the source `Vec` exactly.

Instead, use an efficient implementation which doesn't perform any bounds checks. This simpler implementation is particularly efficient when `T` is `Copy`, as compiler can batch reads and writes into a series of 16-byte SIMD ops.

This should benefit Rolldown, which uses `clone_in_with_semantic_ids` to clone the entirety of AST and `Scoping`. It also gives a +2% speed-up to isolated declarations.