feat(allocator): implement `CloneIn` for `Box<[T]>` (#13505)
Implement `CloneIn` for `Box<[T]>`. Use an efficient implementation which doesn't perform any bounds checks. Hopefully, where `T` is `Copy`, compiler will be able to boil this down to a single `memcpy` call.

(this stack is mostly leftovers from #13450 which turned out not to be required for that PR, but are still useful for other purposes)