feat(semantic): add ability to lookup if AST contains any node kinds (#13137)
I decided to revive this stack: https://github.com/oxc-project/oxc/pull/12227 and merge all of the changes into one PR for slightly easier review and changes.

This PR adds a new `AstTypesBitset` struct which is a set of bits that has a fixed size known at compile time and can be used to store AST types as a set. The aim of this structure is to making checks like "does this file have a for statement?" possible to answer without doing any additional iterations over the AST.