refactor(linter/plugins): use fixed-size allocators when `ExternalLinter` exists (#13623)
We're trying to move the linter away from relying on Cargo features to enable/disable support for JS plugins, and over to runtime options.

The signal that JS plugins are in use is the existence of an `ExternalLinter`.

Select what kind of `AllocatorPool` to use based on this criteria - fixed-size allocators if JS plugins are in use, or standard allocators otherwise. Utilize the `AllocatorPool::new_fixed_size` method added in #13622 to do this.