fix(minifier): disallow merging assignments to let declarations when TDZ error would be introduced (#13635)
```js
let x;
x = (() => x?.a)();
```
should not be transformed to
```js
let x = (() => x?.a)();
```
, as it would introduce a TDZ error.

To fix that, I limited the right hand side of the assignment to be a literal value for let declarations.

The precise check would be to check whether the right hand side includes a reference to the variable declared on the left hand side, but I didn't do that for now as it is complicated. If we implement that, we can improve this condition too.
https://github.com/oxc-project/oxc/blob/5d597f71c06a9c56fe404a712c05b61e2da3697a/crates/oxc_minifier/src/peephole/minimize_statements.rs#L544-L548

fixes #13615
refs #13270