build(napi/oxlint): do not fully minify JS code (#13661)
Alter TSDown settings for `napi/oxlint2`.

We intended the code to be minified, but not to have symbols mangled, or whitespace removed, so it remains somewhat readable. It seems that TSDown may have changed how it handles these options, as `minify: { compress: true }` now turns on mangling and whitespace removal.

Alter the TSDown options to get the intended output.