refactor(linter): add `--experimental-js-plugins` CLI arg (#13658)
Add `--experimental-js-plugins` CLI option to `oxlint`.

At present this option is hidden in the CLI's `--help` output.

If this option is not present, `ExternalLinter` is dropped, which disables JS plugins.

Therefore, JS plugins are only enabled, and code related to them (e.g. fixed size allocators) runs only if both:

1. `oxlint` is run via `napi/oxlint2`.
2. `--experimental-js-plugins` CLI option is present.

This prepares the way for switching `oxlint` over to `napi/oxlint2`.