fix(formatter): incorrectly wrap a parenthesis for `ArrowFunctionExpression` when it has a leading type cast comment (#13683)
Previously, wrongly wrapped a parenthesis for `FunctionBody` because it has a leading type cast comment. The problem we only check whether there is a `(` between the type cast comment and the node, but it actually contains `\n() => (\n`.

In this fix, I added an extra check to ensure only `(` and whitespace can be inside it.
```js
/**
 * @type {object}
 */
() => (
    <div>
        sajdfpoiasdjfpoiasdjfpoiasdjfpoiadsjfpaoisdjfapsdiofjapioisadfaskfaspiofjp
    </div>
);
```