fix(formatter): correct comments printing for import and export (#13707)
This PR aims to remove the special logic for import and export declarations from `get_trailing_comments` and move the special logic to the printing site. The benefits of this are to make `get_trailing_comments` cleaner so that we can entirely remove `SiblingNode`, which may lead to significant performance improvements.