style(linter): add comments + re-organise imports (#13715)
Pure refactor. Add a couple more comments to `apps/oxlint`, and re-order a couple of imports.

This is last bit of refactoring before moving code from `napi/oxlint` into `apps/oxlint`. Have put this in a separate PR to minimize the diff of the NAPI migration PR as much as possible.