test(linter): set CWD for tests (#13722)
#13723 turns `apps/oxlint` into a NAPI package, and moves all the files from `napi/oxlint` into `apps/oxlint`.

These files include the `test` directory, which includes tests for JS plugins. This causes problems with 3 of the existing `oxlint` Rust tests, which run with `cwd` as `apps/oxlint` and no config file specified, because they pick up config files in the JS test fixtures. These config files include JS plugins, which `oxlint`'s Rust tests can't handle because they're run without an `ExternalLinter`.

This PR works around this problem by setting `cwd` to `apps/oxlint/fixtures` for these 3 tests.

I'm not sure if this is a reasonable way to solve the problem.