fix(linter/react/jsx-handler-names): do not detect the function name within the inline-function's body block (#13456)
This PR fixes a behavior of `react/jsx-handler-names` rule incompatible
with that of the eslint-version.
## Fixed difference with the eslint version
With the configuration below:
```json
{ "checkInlineFunction": true, "checkLocalVariables": true }
```
the eslint version of `react/jsx-handler-names` rule treats the
following code as invalid.
```tsx
<TestComponent onChange={() => { handleChange() }} />
```
But the current oxc version finds `handleChange` to be an event handler
name and considers this code to be correct.

Along with this change, the diagnostic messages are also modified
because the event handler name is not always available, and replaced
"invalid" with "bad" to be in accordance with the eslint-version's
message ID`badHandlerName` and `badPropKey`.

---------

Co-authored-by: Cameron Clark <cameron.clark@hey.com>