fix(minifier): don't inline single use variables that are not literals to for statement initializers (#13769)
```js
function wrapper() { var x = a; for (let a of x) console.log(a) }
```
should not be compress to
```js
function wrapper() { for (let a of a) console.log(a) }
```
because `a` points to a different variable as the scope is different.
This PR fixes those kinds of cases for all for statements.

Found in https://github.com/rolldown/rolldown/pull/6220

refs #13755