fix(linter/no-unsafe-declaration-merging): always mark first span as primary (#13830)
This is technically a breaking change, but more closely aligns our implementation with typescript-eslint, and makes placement of the disable directive more predictable.

Previously, it was unclear where the disable directive should go (on the clase or on the interface), this PR enforces that the directive must go on the `class`.