refactor(linter/plugins): make `Context` properties getters (#13809)
Make all properties of `Context` read-only, and accessed via getters.

In ESLint, these properties are made read-only by making them `{ writable: false, configurable: false }`. We can't do that, because the same `Context` object is reused for every file, so it needs to be writable internally. So converting property accesses to getters is the closest we can get to ESLint's behavior.