refactor(language_server): introduce `LSPFileSystem` (#13731)
Formatting requests has no source text included:
https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#documentFormattingParams

Because users can format in memory files, the server needs to keep track of the changes.
Created in `Backend` struct, so when we later support `textDocument/diagnostic` (we support currently only `textDocument/publishDiagnostic)`, we can use the same File System for it too :)

Putting the struct into a Workspace Worker is an option too, but there can be multiple of it.
With this solution, there is only one instance of it :)