fix(build): bump package version in `apps/oxlint` in pre-release workflow (#13895)
Package version is used by NAPI-RS in the bindings file it generates.

Adapt the pre-release workflow to run NAPI-RS build when creating a release PR, so that the version number in bindings files matches the new version number.

This change covers all NAPI packages in `crates` pre-release workflow, and also `oxlint` in its pre-release workflow.