feat(data_structures): add `advance_to_end` method to `SliceIter` trait (#13947)
Add `advance_to_end` method to `SliceIter` trait. This allows calling `advance_to_end` on any slice iterator.

(this stack of 3 PRs is work I did as part of first attempt to build an ESTree walker, which I abandoned, but the APIs are useful for other tasks in `oxc_ast_tools`).