refactor(linter): make disable directives own the rule name (#13987)
doing this obviously has a perf and memory implication. The alternative is that we have to keep the source text in memory until tsgolint has finished linting (and all source files must be in memory the whole time), which is a worse trade off. Now that disable directives struct owns the rule names we can drop the source text (reducing memory usage) while tsgolint runs