perf(linter): add node type codegen support for `match node.kind()` (#13736)
- part of github.com/oxc-project/oxc/issues/12223

Continuation of PRs that generate node types for lint rules. This implements the last main syntax pattern in lint rules which looks like this:

```rust
match node.kind() {
  AstKind::Function(_) => ...,
  AstKind::ArrowFunctionExpression(_) => ...,
  _ => {},
}
```

For this example, we would generate `[Function, ArrowFunctionExpression]`. We do not support guards at the moment, or any non tuple-struct patterns. The wildcard block must be empty, and the match statement must be the only one in the `run` function. This is so that we can be absolutely sure the list of node types is correct, so we tend to be more conservative in what we allow.