feat(linter/plugins): add `defineRule` API (#13945)
Add `defineRule` method for defining rules. This is the 2nd part of the "alternative API" mentioned in https://github.com/oxc-project/oxc/issues/9905#issuecomment-3208424180.

```js
import { defineRule } from 'oxlint';

export default defineRule({
  meta: { /* ... */ },
  create(context) {
    // Rule implementation
  }
});
```

or, using the alternative `createOnce` API:

```js
export default defineRule({
  meta: { /* ... */ },
  createOnce(context) {
    // Rule implementation
  }
});
```

When using `createOnce`, `defineRule` creates a `create` method and adds it to the rule object. So the rule remains ESLint-compatible, while getting a potential speed boost when the rule is used with Oxlint. This PR includes a test that runs an Oxlint plugin that uses `defineRule` + `createOnce` in ESLint.

Not done yet: We need to export TS type defs, so the type checking benefit of `defineRule` is realized. I'll do that in a later PR.

---

`defineRule` is intended to be used when defining JS plugins, so we should move it into a separate NPM package (`oxlint-tools`?), so that someone publishing a plugin to NPM doesn't need their plugin to have a dependency on `oxlint` itself, which includes large binaries.

But for now, for simplicity, just include it in `oxlint` package.