fix(linter/plugins): `defineRule` call `createOnce` lazily (#14062)
Previously `defineRule`, when passed a rule with a `createOnce` method, would call `createOnce` immediately.

When the rule is used in Oxlint, this resulted in `createOnce` being called twice. `createOnce` might do some fairly expensive set-up work, so we only want to call it once (as the name suggests!).

Instead, `defineRule` don't call `createOnce` immediately. Call it the first time the `create` method which `defineRule` creates is called.

This means that, when using Oxlint:

1. `createOnce` is only called once.
2. `defineRule` call is cheaper.

i.e. the cost of making the rule ESLint-compatible is only paid when the user is running the rule in ESLint, not in Oxlint.